fix: UI back button logging and add photo functionality
It was imposible to go back from the loggin activity while the user wasnt logged in using the back button. We fix this through a adding on ActivityResult and handling the realease of resources in onPause and on Resume. In the other hand, we added the functionality of the add photo button.

Resolves:NA
See also:NA